+++ src/main/java/cn/duojunrui/etims/project/module/material/mapper/MaterialMapper.java
@@ -19,7 +19,7 @@ public interface MaterialMapper {
      * @param materialId 教学资源ID
      * @return 教学资源信息
      */
-    public Material selectMaterialById(Long materialId);
+    public Material selectMaterialById(Integer materialId);
 
     /**
      * 查询教学资源列表
